LocalDate should be in the FilingiInfo, not FilingDoc
