+++ src/main/java/edu/suffolk/litlab/efspserver/ecf/TylerModuleSetup.java
@@ -3,7 +3,6 @@ package edu.suffolk.litlab.efspserver.ecf;
 import java.io.IOException;
 import java.sql.SQLException;
 import java.util.HashSet;
-import java.util.List;
 import java.util.Optional;
 import java.util.Set;
